add HiveMessage object add HiveMindNodeType object
