+++ jarbas_hive_mind/interface.py
@@ -1,5 +1,28 @@
 import json
+from jarbas_hive_mind.utils import serialize_message
+from jarbas_hive_mind.message import HiveMessage, HiveMessageType
 from ovos_utils.messagebus import Message
+
+
+def _get_payload(msg):
+    if isinstance(msg, HiveMessage):
+        msg = msg.payload
+    if isinstance(msg, Message):
+        msg = msg.serialize()
+    if isinstance(msg, str):
+        msg = json.loads(msg)
+    return msg
+
+
+def _get_hivemsg(msg):
+    if isinstance(msg, str):
+        msg = json.loads(msg)
+    if isinstance(msg, dict):
+        msg = HiveMessage(**msg)
+    if isinstance(msg, Message):
+        msg = HiveMessage(msg_type=HiveMessageType.BUS, payload=msg)
+    assert isinstance(msg, HiveMessage)
+    return msg
 
 
 class HiveMindSlaveInterface:
@@ -21,73 +44,67 @@ class HiveMindSlaveInterface:
         return self.client.peer
 
     def send(self, payload):
-        if isinstance(payload, dict):
-            payload = json.dumps(payload)
+        payload = serialize_message(payload)
         if not isinstance(payload, bytes):
             payload = bytes(payload, encoding="utf-8")
-
         self.client.sendMessage(payload)
 
     def send_to_hivemind_bus(self, payload):
-        if isinstance(payload, Message):
-            payload = payload.serialize()
-        payload = {"msg_type": "bus",
-                   "payload": payload,
+        payload = {"msg_type": HiveMessageType.BUS,
+                   "payload": _get_payload(payload),
                    "node": self.node_id,
                    "source_peer": self.peer
                    }
         self.send(payload)
 
-    def broadcast(self, payload, msg_data=None):
-        msg_data = msg_data or {}
-        payload = {"msg_type": "broadcast",
-                   "payload": payload,
-                   "route": msg_data.get("route", []),
+    def broadcast(self, message):
+        message = _get_hivemsg(message)
+        payload = {"msg_type": HiveMessageType.BROADCAST,
+                   "payload": message,
+                   "route": message.route,
                    "source_peer": self.peer,
                    "node": self.client.node_id
                    }
         if self.bus:
-            message = Message("hive.send",
-                              payload,
+            message = Message("hive.send", payload,
                               {"destination": "hive",
                                "source": self.peer})
             self.bus.emit(message)
 
-    def propagate(self, payload, msg_data=None):
-        msg_data = msg_data or {}
-        payload = {"msg_type": "propagate",
-                   "payload": payload,
-                   "route": msg_data.get("route", []),
+    def propagate(self, message):
+        message = _get_hivemsg(message)
+        payload = {"msg_type": HiveMessageType.PROPAGATE,
+                   "payload": message,
+                   "route": message.route,
                    "source_peer": self.peer,
                    "node": self.client.node_id
                    }
 
-        no_send = [n["source"] for n in msg_data.get("route", [])]
+        no_send = [n["source"] for n in message.route]
         if self.peer not in no_send:
             self.send(payload)
 
-        if self.bus and msg_data.get("node", "") == \
-                self.peer + ":MASTER":
-            message = Message("hive.send",
-                              payload,
+        if self.bus and payload.get("node", "") == self.peer + ":MASTER":
+            message = Message("hive.send", payload,
                               {"destination": "hive",
                                "source": self.peer})
             self.bus.emit(message)
 
-    def escalate(self, payload, msg_data=None):
-        msg_data = msg_data or {}
-        payload = {"msg_type": "escalate",
-                   "payload": payload,
-                   "route": msg_data.get("route", []),
+    def escalate(self, message):
+        print("esc1:", message)
+        message = _get_hivemsg(message)
+        print("esc2:", message)
+        # when master receives this it will
+        # update "targets" field with it's own id
+        route_data = {"source": self.peer,
+                      "targets": []}
+        message.update_hop_data(route_data)
+        payload = {"msg_type": HiveMessageType.ESCALATE,
+                   "payload": message,
+                   "route": message.route,
                    "source_peer": self.peer,
                    "node": self.client.node_id
                    }
-
-        route_data = {"source": self.peer,
-                      # when master receives this it will update this field
-                      "targets": [self.peer]}
-        if route_data not in payload["route"]:
-            payload["route"].append(route_data)
         self.send(payload)
 
     def query(self, payload, msg_data=None):
@@ -119,12 +136,12 @@ class HiveMindMasterInterface:
         return self.listener.bus
 
     def send(self, payload, client):
+        payload = serialize_message(payload)
         if isinstance(client, str):
             client = self.listener.clients[client]
-        if isinstance(payload, dict):
-            payload = json.dumps(payload)
         if not isinstance(payload, bytes):
             payload = bytes(payload, encoding="utf-8")
+
         client.sendMessage(payload)
 
     def send_to_many(self, payload, no_send=None):
@@ -142,50 +159,46 @@ class HiveMindMasterInterface:
             if client.peer not in no_send:
                 self.send(payload, client)
 
-    def broadcast(self, payload, msg_data=None):
-        msg_data = msg_data or {}
-        payload = {"msg_type": "broadcast",
-                   "payload": payload,
-                   "route": msg_data.get("route", []),
+    def broadcast(self, message):
+        message = _get_hivemsg(message)
+        payload = {"msg_type": HiveMessageType.BROADCAST,
+                   "payload": message,
+                   "route": message.route,
                    "source_peer": self.peer,
                    "node": self.node_id
                    }
-        no_send = [n["source"] for n in msg_data.get("route", [])]
+        no_send = [n["source"] for n in message.route]
         self.send_to_many(payload, no_send)
 
-    def propagate(self, payload, msg_data=None):
-        msg_data = msg_data or {}
-        payload = {"msg_type": "propagate",
-                   "payload": payload,
-                   "route": msg_data.get("route", []),
+    def propagate(self, message):
+        message = _get_hivemsg(message)
+        payload = {"msg_type": HiveMessageType.PROPAGATE,
+                   "payload": message,
+                   "route": message.route,
                    "source_peer": self.peer,
                    "node": self.node_id
                    }
-
-        no_send = [n["source"] for n in msg_data.get("route", [])]
+        no_send = [n["source"] for n in message.route]
         self.send_to_many(payload, no_send)
 
         # tell Slave to propagate upstream
-        if self.bus and msg_data.get("source_peer") != self.peer:
-            payload["msg_type"] = "propagate"
-            message = Message("hive.send",
-                              payload,
+        if self.bus and message.source_peer != self.peer:
+            message = Message("hive.send", payload,
                               {"destination": "hive",
                                "source": self.peer})
             self.bus.emit(message)
 
-    def escalate(self, payload, msg_data=None):
-        msg_data = msg_data or {}
-        payload = {"msg_type": "escalate",
-                   "payload": payload,
-                   "route": msg_data.get("route", []),
+    def escalate(self, message):
+        message = _get_hivemsg(message)
+        payload = {"msg_type": HiveMessageType.ESCALATE,
+                   "payload": message,
+                   "route": message.route,
                    "source_peer": self.peer,
                    "node": self.node_id
                    }
         # tell Slave to escalate upstream
-        if self.bus and msg_data.get("source_peer") != self.peer:
-            message = Message("hive.send",
-                              payload,
+        if self.bus and message.source_peer != self.peer:
+            message = Message("hive.send", payload,
                               {"destination": "hive",
                                "source": self.peer})
             self.bus.emit(message)
