+++ setup.py
@@ -4,17 +4,20 @@ setup(
     name='jarbas_hive_mind',
     version='0.10.7',
     packages=['jarbas_hive_mind',
-              'jarbas_hive_mind.master',
-              'jarbas_hive_mind.slave',
+              'jarbas_hive_mind.nodes',
               'jarbas_hive_mind.configuration',
               'jarbas_hive_mind.database',
               'jarbas_hive_mind.utils',
-              'jarbas_hive_mind.discovery'],
+              'jarbas_hive_mind.discovery',
+               # below are deprecated, backwards compat only
+              'jarbas_hive_mind.master',
+              'jarbas_hive_mind.slave'],
     include_package_data=True,
     install_requires=["pyopenssl",
                       "service_identity",
                       "autobahn",
                       "twisted",
+                      "mycroft-messagebus-client>=0.9.1",
                       "ovos_utils>=0.0.6",
                       "json_database>=0.2.6",
                       "pycryptodomex",
